passing everything to vue

+++ app/Http/Controllers/PostsController.php
@@ -18,27 +18,30 @@ class PostsController extends Controller
         return $posts;
     }
 
-
-    public function create(Request $request){
+    public function storage(Request $request){
         $user_id = Auth::user()->id;
+
         $request->validate($this->rules());
 
         $post = Post::create([
             "user_id" => $user_id,
             "text" => $request["text"]
         ]);
-        $posts = [$post];
-        $view = view('posts', ["posts" => $posts, "auth_id" => $user_id])->render();
-
+       
         //broadcast(new NewPost($post, Auth::user(), $view))->toOthers();
-
-        return response()->json(["html" => $view, "count" => count($posts)]);
+        return response()->json(['post' => $post->where('id', '=', $post->id)->with('user')->first()], 201);
     }
 
     public function destroy(Request $request){
-        $post = (new Post())->find(base64_decode($request["id"]));
-        if(!is_null($post)){
-            $post->delete();
+        if($request->has("id")){
+            $post = (new Post())->find($request["id"]);
+            if(!is_null($post)){
+              $post->delete();
+            }else{
+               return response()->json(['msg' => "Not Found Post With this id"], 404);
+            }
+        }else{
+            return response()->json(['msg' => "Missing param id"], 400);
         }
     }
 
@@ -46,7 +49,6 @@ class PostsController extends Controller
         $rules = [
             "text" => "string|min:1|max:140|required"
         ];
-
 
         return $rules;
     }
